pyproject.toml
[tool.poetry]
name = "abbrbot"
version = "0.1.0"
description = ""
authors = ["Anton Skvortsov <askvrtsv@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
